inline-styles removed
