+++ src/App.tsx
@@ -11,38 +11,9 @@ import SalesAnalytics from './components/SalesAnalytics';
 import SellingProducts from './components/SellingProducts';
 import Sidebar from './components/SideBarComponent/SideBar';
 import Header from './components/HeaderComponent/Header';
+import barchartData from "./Assets/data/BarData"
 
 function App() {
-  const barchartData = [
-    {
-      title: 'Mon',
-      value: 50,
-    },
-    {
-      title: 'Tue',
-      value: 80,
-    },
-    {
-      title: 'Wed',
-      value: 20,
-    },
-    {
-      title: 'Thu',
-      value: 30,
-    },
-    {
-      title: 'Fri',
-      value: 75,
-    },
-    {
-      title: 'Sat',
-      value: 55,
-    },
-    {
-      title: 'Sun',
-      value: 35,
-    },
-  ];
 
   return (
     <>
