+++ src/Assets/data/BarData.tsx
@@ -1,0 +1,32 @@
+const barchartData = [
+    {
+      title: 'Mon',
+      value: 50,
+    },
+    {
+      title: 'Tue',
+      value: 80,
+    },
+    {
+      title: 'Wed',
+      value: 20,
+    },
+    {
+      title: 'Thu',
+      value: 30,
+    },
+    {
+      title: 'Fri',
+      value: 75,
+    },
+    {
+      title: 'Sat',
+      value: 55,
+    },
+    {
+      title: 'Sun',
+      value: 35,
+    },
+  ];
+
+  export default barchartData;
